prepare for Arduino V3.0

+++ yoRadio/src/audioVS1053/audioVS1053Ex.cpp
@@ -5,7 +5,7 @@
  *  vs1053_ext.cpp
  *
  *  Created on: Jul 09.2017
- *  Updated on: Oct 22.2023
+ *  Updated on: Oct 23.2023
  *      Author: Wolle
  */
 #ifndef VS_PATCH_ENABLE
@@ -849,7 +849,7 @@ void Audio::processWebStream() {
 	
     // buffer fill routine - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
     if(availableBytes) {
-        availableBytes = min(availableBytes, InBuff.writeSpace());
+        availableBytes = min(availableBytes, (uint32_t)InBuff.writeSpace());
         int16_t bytesAddedToBuffer = _client->read(InBuff.getWritePtr(), availableBytes);
 
         if(bytesAddedToBuffer > 0) {
